Snapshot considering moving populate/migration to server
